failover/app/requirements.txt
psycopg2-binary
requests
Flask>=0.9
urllib3




